+++ onnx/main.py
@@ -10,6 +10,7 @@ from kivy.core.window import Window
 from kivy.metrics import dp, sp
 from kivy.utils import platform
 from kivy.uix.image import Image
+from kivy.uix.camera import Camera
 
 from kivymd.app import MDApp
 from kivymd.uix.navigationdrawer import MDNavigationDrawerMenu
@@ -23,11 +24,12 @@ Window.softinput_mode = "below_target"
 
 # Import your local screen classes & modules
 from screens.img_obj_detect import ImgObjDetBox, TempSpinWait
+from screens.cam_obj_detect import CamObjDetBox
 from screens.setting import SettingsBox
 from onnx_vision import OnnxDetect
 
 ## Global definitions
-__version__ = "0.0.3"
+__version__ = "0.0.4"
 # Determine the base path for your application's resources
 if getattr(sys, 'frozen', False):
     # Running as a PyInstaller bundle
@@ -49,6 +51,8 @@ class VisionAiApp(MDApp):
     image_path = StringProperty("")
     op_img_path = StringProperty("")
     onnx_detect = ObjectProperty(None)
+    cam_found = ObjectProperty(None)
+    camera = ObjectProperty(None)
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
@@ -57,23 +61,6 @@ class VisionAiApp(MDApp):
     def build(self):
         self.theme_cls.primary_palette = "Blue"
         self.theme_cls.accent_palette = "Orange"
-        self.top_menu_items = {
-            "Documentation": {
-                "icon": "file-document-check",
-                "action": "web",
-                "url": "https://blog.daslearning.in/llm_ai/genai/image-to-animation.html",
-            },
-            "Contact Us": {
-                "icon": "card-account-phone",
-                "action": "web",
-                "url": "https://daslearning.in/contact/",
-            },
-            "Check for update": {
-                "icon": "github",
-                "action": "update",
-                "url": "",
-            }
-        }
         return Builder.load_file(kv_file_path)
 
     def on_start(self):
@@ -81,10 +68,21 @@ class VisionAiApp(MDApp):
         if platform == "android":
             from android.permissions import request_permissions, Permission
             from jnius import autoclass, PythonJavaClass, java_method
-            request_permissions([
-                Permission.READ_EXTERNAL_STORAGE,
-                Permission.WRITE_EXTERNAL_STORAGE
-            ])
+            sdk_version = 28
+            try:
+                VERSION = autoclass('android.os.Build$VERSION')
+                sdk_version = VERSION.SDK_INT
+                print(f"Android SDK: {sdk_version}")
+                #self.show_toast_msg(f"Android SDK: {sdk_version}")
+            except Exception as e:
+                print(f"Could not check the android SDK version: {e}")
+                #self.show_toast_msg(f"Error checking SDK: {e}", is_error=True)
+            permissions = [Permission.CAMERA]
+            if sdk_version >= 33:  # Android 13+
+                permissions.append(Permission.READ_MEDIA_IMAGES)
+            else:  # Android 10–12
+                permissions.extend([Permission.READ_EXTERNAL_STORAGE, Permission.WRITE_EXTERNAL_STORAGE])
+            request_permissions(permissions)
             context = autoclass('org.kivy.android.PythonActivity').mActivity
             android_path = context.getExternalFilesDir(None).getAbsolutePath()
             self.model_dir = os.path.join(android_path, 'model_files')
@@ -98,8 +96,9 @@ class VisionAiApp(MDApp):
             except Exception:
                 self.external_storage = os.path.abspath("/storage/emulated/0/")
         else:
-            self.internal_storage = os.path.abspath("/")
-            self.external_storage = os.path.abspath("/")
+            self.internal_storage = os.path.expanduser("~")
+            self.external_storage = os.path.expanduser("~")
+            os.makedirs(os.path.join(self.internal_storage, 'images'), exist_ok=True)
             self.model_dir = os.path.join(self.user_data_dir, 'model_files')
             self.op_dir = os.path.join(self.user_data_dir, 'outputs')
         os.makedirs(self.model_dir, exist_ok=True)
@@ -156,8 +155,47 @@ class VisionAiApp(MDApp):
             self.txt_dialog.dismiss()
 
     def on_img_obj_detect(self):
-        print(f"Entered the image object detection screen")
         self.show_toast_msg("You can detect objects on your image!")
+
+    def on_cam_obj_detect(self):
+        self.show_toast_msg("Capture an image & detect objects!")
+        self.cam_uix = self.root.ids.cam_detect_box.ids.capture_image
+        self.cam_uix.clear_widgets()
+        if platform == "android":
+            cam_indx = 0
+            resolution = (960, 720) # will fallback to 480 if fails again
+        else:
+            resolution = (640, 480)
+            import cv2
+            available_cameras = []
+            for i in range(3):
+                cap = cv2.VideoCapture(i)
+                if cap.isOpened():
+                    print(f"Camera found at index: {i}")
+                    available_cameras.append(i)
+                    cap.release()
+            if len(available_cameras) >= 1:
+                cam_indx = available_cameras[0]
+            else:
+                self.show_toast_msg(f"No camera found on {platform}!", is_error=True)
+                return
+        try:
+            self.camera = Camera(
+                index = cam_indx,
+                resolution = resolution,
+                fit_mode = "contain",
+                play = True
+            )
+            self.cam_uix.add_widget(self.camera)
+            self.cam_found = True
+        except Exception as e:
+            print(f"Error setting up the camera: {e}")
+            self.show_toast_msg(f"Error setting up the camera: {e}", is_error=True)
+
+    def on_cam_obj_dt_leave(self):
+        if self.cam_found:
+            self.camera.play = False
+            self.cam_uix.clear_widgets()
 
     def open_img_file_manager(self):
         """Open the file manager to select an image file. On android use Downloads or Pictures folders only"""
@@ -218,7 +256,10 @@ class VisionAiApp(MDApp):
         download_stat = self.download_n_remove_file(self.op_img_path, chosen_path)
         if download_stat:
             self.op_img_path = ""
-            result_box = self.root.ids.img_detect_box.ids.result_image
+            if self.root.ids.screen_manager.current == "imgObjDetect":
+                result_box = self.root.ids.img_detect_box.ids.result_image
+            else:
+                result_box = self.root.ids.cam_detect_box.ids.cam_result_image
             result_box.clear_widgets()
         self.op_file_exit_manager()
 
@@ -234,7 +275,7 @@ class VisionAiApp(MDApp):
         if self.is_onnx_running:
             self.show_toast_msg("Please wait for the previous request to finish", is_error=True)
             return
-        onnx_thread = Thread(target=self.onnx_detect.run_detect, args=(self.image_path, self.onnx_detect_callback), daemon=True)
+        onnx_thread = Thread(target=self.onnx_detect.run_detect, args=(self.image_path, self.onnx_detect_callback, "imgObjDetect"), daemon=True)
         onnx_thread.start()
         self.is_onnx_running = True
         tmp_spin = TempSpinWait()
@@ -242,14 +283,42 @@ class VisionAiApp(MDApp):
         result_box.clear_widgets()
         result_box.add_widget(tmp_spin)
 
+    def capture_n_onnx_detect(self):
+        if not self.cam_found:
+            self.show_toast_msg("Camera could not be loaded!", is_error=True)
+            return
+        if self.is_onnx_running:
+            self.show_toast_msg("Please wait for the previous request to finish", is_error=True)
+            return
+        self.image_path = ""
+        import datetime
+        now = datetime.datetime.now()
+        current_time = str(now.strftime("%H%M%S"))
+        current_date = str(now.strftime("%Y%m%d"))
+        internal_dir = os.path.join(self.internal_storage, 'images')
+        capture_file = f"aivision-{current_date}-{current_time}.png"
+        self.image_path = os.path.join(internal_dir, capture_file)
+        self.camera.export_to_png(self.image_path)
+        onnx_thread = Thread(target=self.onnx_detect.run_detect, args=(self.image_path, self.onnx_detect_callback, "camObjDetect"), daemon=True)
+        onnx_thread.start()
+        self.is_onnx_running = True
+        tmp_spin = TempSpinWait()
+        result_box = self.root.ids.cam_detect_box.ids.cam_result_image
+        result_box.clear_widgets()
+        result_box.add_widget(tmp_spin)
+
     def onnx_detect_callback(self, onnx_resp):
         status = onnx_resp["status"]
         message = onnx_resp["message"]
+        caller = onnx_resp["caller"]
         self.is_onnx_running = False
+        if caller == "camObjDetect":
+            result_box = self.root.ids.cam_detect_box.ids.cam_result_image
+        else:
+            result_box = self.root.ids.img_detect_box.ids.result_image
         if status is True:
             self.show_toast_msg(f"Output generated at: {message}")
             self.op_img_path = message
-            result_box = self.root.ids.img_detect_box.ids.result_image
             result_box.clear_widgets()
             fitImage = Image(
                 source = message,
@@ -273,6 +342,11 @@ class VisionAiApp(MDApp):
         uploaded_image_box = self.root.ids.img_detect_box.ids.uploaded_image
         uploaded_image_box.clear_widgets()
         result_box = self.root.ids.img_detect_box.ids.result_image
+        result_box.clear_widgets()
+
+    def reset_cam_object_detect(self):
+        self.image_path = ""
+        result_box = self.root.ids.cam_detect_box.ids.cam_result_image
         result_box.clear_widgets()
 
     def open_link(self, instance, url):
@@ -352,6 +426,14 @@ class VisionAiApp(MDApp):
                 else:
                     self.img_file_manager.back()  # Navigate back within file manager
                 return True  # Consume the event to prevent app exit
+            if self.is_op_file_mgr_open:
+                # Check if we are at the root of the directory tree
+                if self.op_file_manager.current_path == self.external_storage:
+                    self.show_toast_msg(f"Closing file manager from main storage")
+                    self.op_file_exit_manager()
+                else:
+                    self.op_file_manager.back()  # Navigate back within file manager
+                return True
         return False
 
 if __name__ == '__main__':